fix(#1180): honor the printf/scanf %L length modifier
The VM's formatter dropped the `L` length modifier: `printf("%Lf", x)`
emitted the literal "Lf" and consumed no argument, desyncing every later
conversion. Three distinct defects on three platform/path combinations,
only the first of which the ticket named:

1. stb_sprintf.h's size-override switch had no `case 'L'`, so `L` fell
   through to the "unknown conversion, copy literally" default (macOS,
   glibc < 2.35).
2. On glibc >= 2.35 guest printf binds straight to host printf, which
   pops a real 16-byte x87 long double while the VM pushed an 8-byte
   ffi_type_double -- a TY_LDOUBLE variadic arg was always marshalled as
   ffi_type_double.
3. format_scanf.c's store_float wrote a host long double through the
   guest pointer for %Lf; the VM reads that object back as a flat 8-byte
   double (fine on macOS arm64, garbage on both Linux targets).

Fix: marshal a variadic long double as a real ffi_type_longdouble at the
FFI boundary, encoded as BOTH arg-mask bits set (an otherwise-impossible
combination, so no bytecode-encoding change), widening the VM's double
there. stb_sprintf.h and the v* re-dispatch helper read the wider slot to
match; store_float stores a flat double. Variadic-only -- a long double
in a FIXED parameter keeps the plain double binding, since every ...l
libc symbol is a double-precision shim (#491). The %L conversions
themselves still format at double precision.

Deliberate consequence: a long double passed to a plain %f/%e/%g is now a
genuine argument-width mismatch. The -F format checker flags it
(FMT_EXPECT_DOUBLE no longer accepts TY_LDOUBLE), matching gcc/clang
-Wformat.

New coverage in tests/suites/test_suite_floats.c: %Lf/%Le/%Lg/%LF/%La,
a slot-desync guard, the v* forwarder path, sscanf %Lf, and a sqrtl
fixed-parameter guard.

Verified: full suite + --native (786/786) + comptime_native_smoke +
--build-target=test on macOS arm64, and the full sharded suite in the
linux/amd64 (x86-64, 80-bit long double) container.

+++ src/stdlib/format_scanf.c
@@ -121,7 +121,12 @@ static void store_int(void *ptr, int lenmod, unsigned long long val) {
 
 static void store_float(void *ptr, int lenmod, long double val) {
     if (lenmod == LEN_L)
-        *(long double *)ptr = val;
+        // #1180: the guest's `long double` object is an 8-byte flat double
+        // (the VM reads it back with a plain FLDR; CCCC models long double
+        // arithmetic at double precision, #491). Storing a host `long
+        // double` here -- 80-bit x87 on Linux/x86-64, binary128 on
+        // Linux/aarch64 -- would write a representation the VM cannot read.
+        *(double *)ptr = (double)val;
     else if (lenmod == LEN_l)
         *(double *)ptr = (double)val;
     else
